Split streams init into c file
